% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrMNAR.R
\name{msIpwQr}
\alias{msIpwQr}
\title{Estimates quantile regression coefficients with nonignorable missing data}
\usage{
msIpwQr(
  sim_data,
  p1,
  p2,
  q1,
  q2,
  thresh,
  max_iter.ms,
  tau_seq.out,
  tau_seq.ms,
  m
)
}
\arguments{
\item{sim_data}{imported data frame of the form sim_data = data.frame(X,Y).
the last column of sim_data is the respones Y which subject to missing (NA if missing).
the other columns of sim_data are the covariates X = (X1, X2)= (X1.cat, X1.cont, X2.cat, X2.cont) in order, where
X1 refers to covariates which are related to the response and also included in the misssing mechanism model;
X2 refers to the auxiliary variable, which means given (X1, Y), X2 is not included in the misssing mechanism model;
X1.cat and X2.cat are the categorical variables in X1 and X2, respectively;
X1.cont and X2.cont are the continuous variables in X1 and X2, respectively.}

\item{p1}{the number of categorical variables contained in X1}

\item{p2}{the number of continuous variables contained in X1; p1>0 since X1.cont Include intercept term}

\item{q1}{the number of categorical variables contained in X2}

\item{q2}{the number of continuous variables contained in X2; q1+q2>0 since X2 has a dimension of at least 1}

\item{thresh}{the threshold of this algorithm for convergence}

\item{max_iter.ms}{The maximum iterations of this algorithm;
When the algorithm reaches the maximum number of iterations or threshold, the algorithm stops iterating}

\item{tau_seq.out}{a vector; the quantiles at which the regression coefficients to be estimated for the output
The result will output the coefficient estimates of the quantile regression of these quantile points}

\item{tau_seq.ms}{a vector; the quantiles at which the regression coefficients to be estimated for f(y|x)
when estimating theta, similar as Wei. et al(2009) to estimate the density f(y|x)}

\item{m}{the number of samples drawn from f(y|x) for the missing response Y}
}
\value{
the output is a list which include seven elements;
 the first element of the output is a vector, including the number of iterations and the maximum
 difference of the estimates between the previous and last iterations when the algorithm terminates;
 the second element of the output is a vector, the estimates of missing mechanism coefficients;
 the third element of the output is a matrix, the estimates of quantile regression coefficients at quantile levels "tau_seq.out"
 the fourth element of the output is a matrix, the estimates of missing mechanism coefficients for each iteration
 the fifth element of the output is a matrix, the estimates of quantile regression coefficients for each iteration at quantile levels "tau_seq.out".
 the sixth element of the output is a list, the summary of missing mechanism model,
 more details of missing mechanism model can be found here, including p-values which can be used to determine which variables are significant
 the seventh element of the output is a list, the summary of quantile regression model at quantile levels "tau_seq.out".
 more details of missing mechanism model can be found here, including p-values which can be used to determine which variables are significant
}
\description{
This function is used to get the proposed estimates of quantile regression coefficients for nonignorable missing data
}
\examples{
## generate sim_data for example
nsize <- 1000
beta <- c( 1,-2,2,0.5 ); gam <- c( 0.5, 0.5, 0.2, 0); theta <- c(-2, 0.5,0.5,0.5)
x1 <- rbinom( nsize, 1, 0.5); x2 <- rnorm( nsize, mean = 2, sd = 0.5); x3 <- runif( nsize, min = 0, max = 2)
xx <- cbind( 1, x1, x2, x3); err <- xx\%*\%gam*rnorm( nsize, mean = 0, sd = 0.5); yy <- xx \%*\% beta + err
xy <- cbind( 1, x1, x2, yy); prob.m <- 1/( 1+ exp( - xy \%*\% theta ))
delta.ind <- sapply( prob.m, function(p){rbinom(n=1, size=1, prob=p)}); yy[delta.ind==0] <- NA
sim_data <- data.frame( x1, x2, x3, yy)
p1<- 1; p2<- 1; q1<- 0; q2 <- 1

## set the initial value related to the algorithm
thresh = 1e-04; max_iter.ms <- 50;  m <- 100
tau_seq.ms  <- seq(0.05, 0.95, by = 0.05); tau_seq.out <- seq(0.25, 0.75, by = 0.25)

## using sim_data to estimate quantile coefficients
out <- msIpwQr(sim_data, p1, p2, q1, q2, thresh, max_iter.ms, tau_seq.out, tau_seq.ms, m )
beta_tau_true <- beta + gam \%*\% t(qnorm( tau_seq.out, mean = 0, sd = 0.5)) # true value of beta_tau
out[[3]] # quantile coefficients estimates
out[[7]] # summary of quantile regression model
}
